Stop generating NNS .did files, like we did for registry. (Instead of that, assert that implemented interface exactly matches declared interface.)
