+++ rs/nns/governance/canister/canister.rs
@@ -1,14 +1,3 @@
-// Note on `candid_method`: each canister method should have a function
-// annotated with `#[candid_method]` that has the arguments and return type
-// expected by the canister method, to be able to generate `governance.did`
-// automatically.
-//
-// This often means we need a function with `#[export_name = "canister_query
-// my_method"]` that doesn't take arguments and doesn't return anything (per IC
-// spec), then another function the actual method arguments and return type,
-// annotated with `#[candid_method(query/update)]` to be able to generate the
-// did definition of the method.
-
 use async_trait::async_trait;
 use candid::{candid_method, Decode, Encode};
 use dfn_candid::{candid, candid_one};
@@ -1202,129 +1191,9 @@ fn expose_candid() {
     })
 }
 
-// When run on native this prints the candid service definition of this
-// canister, from the methods annotated with `candid_method` above.
-//
-// Note that `cargo test` calls `main`, and `export_service` (which defines
-// `__export_service` in the current scope) needs to be called exactly once. So
-// in addition to `not(target_arch = "wasm32")` we have a `not(test)` guard here
-// to avoid calling `export_service`, which we need to call in the test below.
-#[cfg(not(any(target_arch = "wasm32", test)))]
 fn main() {
-    // The line below generates did types and service definition from the
-    // methods annotated with `candid_method` above. The definition is then
-    // obtained with `__export_service()`.
-    candid::export_service!();
-    std::print!("{}", __export_service());
+    // This block is intentionally left blank.
 }
 
-#[cfg(any(target_arch = "wasm32", test))]
-fn main() {}
-
-#[cfg(not(feature = "test"))]
-#[test]
-fn check_governance_candid_file() {
-    let did_path = std::path::PathBuf::from(
-        std::env::var_os("CARGO_MANIFEST_DIR").expect("CARGO_MANIFEST_DIR env var undefined"),
-    )
-    .join("canister/governance.did");
-    let did_contents = String::from_utf8(std::fs::read(did_path).unwrap()).unwrap();
-
-    // See comments in main above
-    candid::export_service!();
-    let expected = __export_service();
-
-    if did_contents != expected {
-        panic!(
-            "Generated candid definition does not match canister/governance.did. \
-            Run `bazel run :generate_did > canister/governance.did` (no nix and/or direnv) in \
-            rs/nns/governance to update canister/governance.did."
-        )
-    }
-}
-
-#[cfg(feature = "test")]
-#[test]
-fn check_governance_candid_file() {
-    let did_path = std::path::PathBuf::from(
-        std::env::var_os("CARGO_MANIFEST_DIR").expect("CARGO_MANIFEST_DIR env var undefined"),
-    )
-    .join("canister/governance_test.did");
-    let did_contents = String::from_utf8(std::fs::read(did_path).unwrap()).unwrap();
-
-    // See comments in main above
-    candid::export_service!();
-    let expected = __export_service();
-
-    if did_contents != expected {
-        panic!(
-            "Generated candid definition does not match canister/governance_test.did. \
-            Run `bazel run :generate_test_did > canister/governance_test.did` (no nix and/or direnv) in \
-            rs/nns/governance to update canister/governance_test.did."
-        )
-    }
-}
-
-#[test]
-fn test_set_time_warp() {
-    let mut environment = CanisterEnv::new();
-
-    let start = environment.now();
-    environment.set_time_warp(GovTimeWarp { delta_s: 1_000 });
-    let delta_s = environment.now() - start;
-
-    assert!(delta_s >= 1000, "delta_s = {}", delta_s);
-    assert!(delta_s < 1005, "delta_s = {}", delta_s);
-}
-
-#[test]
-fn test_get_effective_payload_sets_proposal_id_for_add_wasm() {
-    let mt = gov_pb::NnsFunction::AddSnsWasm;
-    let proposal_id = 42;
-    let wasm = vec![1, 2, 3];
-    let canister_type = 3;
-    let hash = vec![1, 2, 3, 4];
-    let payload = Encode!(&AddWasmRequest {
-        wasm: Some(SnsWasm {
-            proposal_id: None,
-            wasm: wasm.clone(),
-            canister_type,
-        }),
-        hash: hash.clone(),
-    })
-    .unwrap();
-
-    let effective_payload = get_effective_payload(mt, &payload, proposal_id, 0).unwrap();
-
-    let decoded = Decode!(&effective_payload, AddWasmRequest).unwrap();
-    assert_eq!(
-        decoded,
-        AddWasmRequest {
-            wasm: Some(SnsWasm {
-                proposal_id: Some(proposal_id), // The proposal_id should be set
-                wasm,
-                canister_type
-            }),
-            hash
-        }
-    );
-}
-
-#[test]
-fn test_get_effective_payload_overrides_proposal_id_for_add_wasm() {
-    let mt = gov_pb::NnsFunction::AddSnsWasm;
-    let proposal_id = 42;
-    let payload = Encode!(&AddWasmRequest {
-        wasm: Some(SnsWasm {
-            proposal_id: Some(proposal_id - 1),
-            ..SnsWasm::default()
-        }),
-        ..AddWasmRequest::default()
-    })
-    .unwrap();
-
-    let effective_payload = get_effective_payload(mt, &payload, proposal_id, 0).unwrap();
-
-    let decoded = Decode!(&effective_payload, AddWasmRequest).unwrap();
-    assert_eq!(decoded.wasm.unwrap().proposal_id.unwrap(), proposal_id);
-}
+#[cfg(test)]
+mod tests;
