+++ rs/nns/handlers/root/impl/canister/root.did
@@ -6,9 +6,22 @@ type AddCanisterRequest = record {
   memory_allocation : opt nat;
   compute_allocation : opt nat;
 };
-type CanisterAction = variant { Start; Stop };
-type CanisterIdRecord = record { canister_id : principal };
-type CanisterInstallMode = variant { reinstall; upgrade; install };
+
+type CanisterAction = variant {
+  Start;
+  Stop;
+};
+
+type CanisterIdRecord = record {
+  canister_id : principal;
+};
+
+type CanisterInstallMode = variant {
+  reinstall;
+  upgrade;
+  install;
+};
+
 type CanisterSettings = record {
   freezing_threshold : opt nat;
   controllers : opt vec principal;
@@ -18,6 +31,7 @@ type CanisterSettings = record {
   memory_allocation : opt nat;
   compute_allocation : opt nat;
 };
+
 type CanisterStatusResult = record {
   status : CanisterStatusType;
   memory_size : nat;
@@ -27,22 +41,32 @@ type CanisterStatusResult = record {
   module_hash : opt blob;
   reserved_cycles : opt nat;
 };
-type CanisterStatusType = variant { stopped; stopping; running };
+
+type CanisterStatusType = variant {
+  stopped;
+  stopping;
+  running;
+};
+
 type ChangeCanisterControllersError = record {
   code : opt int32;
   description : text;
 };
+
 type ChangeCanisterControllersRequest = record {
   target_canister_id : principal;
   new_controllers : vec principal;
 };
+
 type ChangeCanisterControllersResponse = record {
   change_canister_controllers_result : ChangeCanisterControllersResult;
 };
+
 type ChangeCanisterControllersResult = variant {
   Ok;
   Err : ChangeCanisterControllersError;
 };
+
 type ChangeCanisterRequest = record {
   arg : blob;
   wasm_module : blob;
@@ -52,6 +76,7 @@ type ChangeCanisterRequest = record {
   memory_allocation : opt nat;
   compute_allocation : opt nat;
 };
+
 type DefiniteCanisterSettings = record {
   freezing_threshold : opt nat;
   controllers : vec principal;
@@ -61,23 +86,32 @@ type DefiniteCanisterSettings = record {
   memory_allocation : opt nat;
   compute_allocation : opt nat;
 };
-type LogVisibility = variant { controllers; public };
+
+type LogVisibility = variant {
+  controllers;
+  public;
+};
+
 type StopOrStartCanisterRequest = record {
   action : CanisterAction;
   canister_id : principal;
 };
+
 type UpdateCanisterSettingsError = record {
   code : opt int32;
   description : text;
 };
+
 type UpdateCanisterSettingsRequest = record {
   canister_id : principal;
   settings : CanisterSettings;
 };
+
 type UpdateCanisterSettingsResponse = variant {
   Ok;
   Err : UpdateCanisterSettingsError;
 };
+
 service : {
   add_nns_canister : (AddCanisterRequest) -> ();
   canister_status : (CanisterIdRecord) -> (CanisterStatusResult);
