+++ rs/nns/sns-wasm/BUILD.bazel
@@ -1,4 +1,4 @@
-load("@rules_rust//rust:defs.bzl", "rust_binary", "rust_library", "rust_test")
+load("@rules_rust//rust:defs.bzl", "rust_library", "rust_test")
 load("//bazel:canisters.bzl", "rust_canister")
 load("//bazel:defs.bzl", "rust_ic_test_suite_with_extra_srcs")
 load("//bazel:prost.bzl", "generated_files_check")
@@ -119,14 +119,6 @@ rust_library(
     deps = DEPENDENCIES_WITH_TEST_FEATURES,
 )
 
-rust_binary(
-    name = "generate_did",
-    srcs = ["canister/canister.rs"],
-    compile_data = ["canister/sns-wasm.did"],
-    proc_macro_deps = MACRO_DEPENDENCIES,
-    deps = DEPENDENCIES + [":sns-wasm"],
-)
-
 rust_canister(
     name = "sns-wasm-canister",
     srcs = ["canister/canister.rs"],
@@ -139,9 +131,11 @@ rust_canister(
 
 rust_test(
     name = "canister_unit_test",
-    crate = ":_wasm_sns-wasm-canister",
-    data = ["canister/sns-wasm.did"],
-    env = {"CARGO_MANIFEST_DIR": "rs/nns/sns-wasm"},
+    crate_root = "canister/canister.rs",
+    srcs = glob(["canister/**/*.rs"]),
+    deps = [":sns-wasm"] + DEPENDENCIES + DEV_DEPENDENCIES,
+    proc_macro_deps = MACRO_DEPENDENCIES,
+    compile_data = ["canister/sns-wasm.did"],
 )
 
 rust_test(
