+++ rs/nns/sns-wasm/canister/canister.rs
@@ -585,6 +585,9 @@ fn http_request() {
     dfn_http_metrics::serve_metrics(encode_metrics);
 }
 
+/// Deprecated: The right way to get this information is to do (the equivalent of)
+/// `dfx canister metadata $CANISTER 'candid:service'`.
+///
 /// This makes this Candid service self-describing, so that for example Candid
 /// UI, but also other tools, can seamlessly integrate with it.
 /// The concrete interface (__get_candid_interface_tmp_hack) is provisional, but
@@ -598,45 +601,10 @@ fn expose_candid() {
     over(candid, |_: ()| include_str!("sns-wasm.did").to_string())
 }
 
-/// When run on native, this prints the candid service definition of this
-/// canister, from the methods annotated with `candid_method` above.
-///
-/// Note that `cargo test` calls `main`, and `export_service` (which defines
-/// `__export_service` in the current scope) needs to be called exactly once. So
-/// in addition to `not(target_arch = "wasm32")` we have a `not(test)` guard here
-/// to avoid calling `export_service`, which we need to call in the test below.
-#[cfg(not(any(target_arch = "wasm32", test)))]
 fn main() {
-    // The line below generates did types and service definition from the
-    // methods annotated with `candid_method` above. The definition is then
-    // obtained with `__export_service()`.
-    candid::export_service!();
-    std::print!("{}", __export_service());
+    // This block is intentionally left blank.
 }
 
-#[cfg(any(target_arch = "wasm32", test))]
-fn main() {}
-
-/// A test that fails if the API was updated but the candid definition was not.
-#[test]
-fn check_wasm_candid_file() {
-    let did_path = std::path::PathBuf::from(
-        std::env::var_os("CARGO_MANIFEST_DIR").expect("CARGO_MANIFEST_DIR env var undefined"),
-    )
-    .join("canister/sns-wasm.did");
-
-    let did_contents = String::from_utf8(std::fs::read(did_path).unwrap()).unwrap();
-
-    // See comments in main above
-    candid::export_service!();
-    let expected = __export_service();
-
-    if did_contents != expected {
-        panic!(
-            "Generated candid definition does not match canister/sns-wasm.did. \
-            Run `bazel run :generate_did > canister/sns-wasm.did` (no nix and/or direnv) or \
-            `cargo run --bin sns-wasm-canister > canister/sns-wasm.did` in \
-            rs/nns/sns-wasm to update canister/sns-wasm.did."
-        )
-    }
-}
+// In order for some of the test(s) within this mod, this MUST occur at the end.
+#[cfg(test)]
+mod tests;
