+++ rs/registry/canister/canister/canister.rs
@@ -1057,25 +1057,26 @@ fn http_request() {
     dfn_http_metrics::serve_metrics(encode_metrics);
 }
 
-// This makes this Candid service self-describing, so that for example Candid
-// UI, but also other tools, can seamlessly integrate with it.
-// The concrete interface (__get_candid_interface_tmp_hack) is provisional, but
-// works.
-//
-// We include the .did file as committed, as means it is included verbatim in
-// the .wasm; using `candid::export_service` here would involve unnecessary
-// runtime computation
-
+/// Deprecated: The right way to get this information is to do (the equivalent of)
+/// `dfx canister metadata $CANISTER 'candid:service'`.
+///
+/// This makes this Candid service self-describing, so that for example Candid
+/// UI, but also other tools, can seamlessly integrate with it.
+/// The concrete interface (__get_candid_interface_tmp_hack) is provisional, but
+/// works.
+///
+/// We include the .did file as committed, as means it is included verbatim in
+/// the .wasm; using `candid::export_service` here would involve unnecessary
+/// runtime computation
 #[export_name = "canister_query __get_candid_interface_tmp_hack"]
 fn expose_candid() {
     over(candid, |_: ()| include_str!("registry.did").to_string())
 }
 
-/// Currently (May, 2024), unlike our other canisters, registry.did is NOT
-/// automatically generated by main. Instead, it is hand-crafted. We might later
-/// change our other canisters to do the same thing. registry is a trial balloon
-/// of sorts in this regard.
-fn main() {}
+fn main() {
+    // This block is intentionally left blank.
+}
 
+// In order for some of the test(s) within this mod, this MUST occur at the end.
 #[cfg(test)]
 mod tests;
